chore: fix composer version dependency for nightly (#97)
Nightly builds initialize the required php infra with composer create-project composer/satis:dev-main.  
This pulls in composer 2.4.2 as a dependency at the time of writing.  
However, the result is a broken upstream nightly and mageos nightly build.
When composer install is run while determining the base package dependencies. The error is:

    - laminas/laminas-dependency-plugin 2.5.0 requires composer-plugin-api >=1.1.0 <2.3.0
           -> found composer-plugin-api[2.3.0] but it does not match the constraint.

Reference
* [failed job](https://github.com/mage-os/generate-mirror-repo-js/actions/runs/3508445982/jobs/5876852570)
* [checkout magento action](https://github.com/mage-os/generate-mirror-repo-js/blob/main/.github/actions/checkout-magento/action.yml#L41)
* [shivammathur/setup-php tools input](https://github.com/shivammathur/setup-php#wrench-tools-support)